+++ json_annotation/lib/src/json_serializable.dart
@@ -89,6 +89,24 @@ class JsonSerializable {
   /// such as [fieldRename].
   final bool? createFieldMap;
 
+  /// If `true` (defaults to false), a private class `_$ExampleJsonKeys`
+  /// class is created in the generated part file.
+  ///
+  /// This class will contain every property as a [String] field with the JSON
+  /// key as the value.
+  ///
+  /// ```dart
+  /// @JsonSerializable(createJsonKeys: true)
+  /// class Example {
+  ///   @JsonKey(name: 'LAST_NAME')
+  ///   String? firstName;
+  ///
+  ///   // Will have the value `LAST_NAME`
+  ///   static const firstName = _$ExampleJsonKeys.firstName;
+  /// }
+  /// ```
+  final bool? createJsonKeys;
+
   /// If `true` (defaults to false), a private, static `_$ExamplePerFieldToJson`
   /// abstract class will be generated in the part file.
   ///
@@ -168,7 +186,7 @@ class JsonSerializable {
   ///   T Function(Object json) fromJsonT,
   /// ) {
   ///   return Response<T>()
-  ///     ..status = json['status'] as int
+  ///     ..status = (json['status'] as num).toInt()
   ///     ..value = fromJsonT(json['value']);
   /// }
   ///
@@ -252,6 +270,7 @@ class JsonSerializable {
     this.checked,
     this.constructor,
     this.createFieldMap,
+    this.createJsonKeys,
     this.createFactory,
     this.createToJson,
     this.disallowUnrecognizedKeys,