+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -54,6 +54,7 @@ class ClassConfig {
   final bool createFactory;
   final bool createToJson;
   final bool createFieldMap;
+  final bool createJsonKeys;
   final bool createPerFieldToJson;
   final bool disallowUnrecognizedKeys;
   final bool explicitToJson;
@@ -73,6 +74,7 @@ class ClassConfig {
     required this.createFactory,
     required this.createToJson,
     required this.createFieldMap,
+    required this.createJsonKeys,
     required this.createPerFieldToJson,
     required this.disallowUnrecognizedKeys,
     required this.explicitToJson,
@@ -95,6 +97,8 @@ class ClassConfig {
         constructor: config.constructor ?? ClassConfig.defaults.constructor,
         createFieldMap:
             config.createFieldMap ?? ClassConfig.defaults.createFieldMap,
+        createJsonKeys:
+            config.createJsonKeys ?? ClassConfig.defaults.createJsonKeys,
         createPerFieldToJson: config.createPerFieldToJson ??
             ClassConfig.defaults.createPerFieldToJson,
         createFactory:
@@ -125,6 +129,7 @@ class ClassConfig {
     createFactory: true,
     createToJson: true,
     createFieldMap: false,
+    createJsonKeys: false,
     createPerFieldToJson: false,
     disallowUnrecognizedKeys: false,
     explicitToJson: false,
@@ -143,6 +148,7 @@ class ClassConfig {
         createFactory: createFactory,
         createToJson: createToJson,
         createFieldMap: createFieldMap,
+        createJsonKeys: createJsonKeys,
         createPerFieldToJson: createPerFieldToJson,
         ignoreUnannotated: ignoreUnannotated,
         explicitToJson: explicitToJson,