+++ json_serializable/lib/src/type_helpers/record_helper.dart
@@ -93,7 +93,7 @@ $helperName(
       );
     }
 
-    final mapValue = '{${items.map((e) => '$e,').join()}}';
+    final mapValue = '<String, dynamic>{${items.map((e) => '$e,').join()}}';
 
     return targetType.isNullableType
         ? ifNullOrElse(